chore: Handle unreachable code from upstream `derive`
`derive(NetworkBehaviour)` expands to code with a dead match arm. Factor
out all structs from struct with the offending derive to minimize the
impact of the added `allow` linting whitelist.

Uses `allow` because the lint rule is only broken with rustc v1.100+,
not on current stable, which is v1.98.
