+++ neptune-core/src/application/network/actor.rs
@@ -38,8 +38,8 @@ use crate::application::network::gateway::StreamGateway;
 use crate::application::network::overview::NetworkOverview;
 use crate::application::network::reachability::ReachabilityState;
 use crate::application::network::stack::NetworkStack;
-use crate::application::network::stack::NetworkStackEvent;
 use crate::application::network::stack::NEPTUNE_PROTOCOL_STR;
+use crate::application::network::stack_event::NetworkStackEvent;
 use crate::macros::fn_name;
 use crate::macros::log_slow_scope;
 use crate::state::GlobalStateLock;
